Add kustomization and labels
A Deployment requires a selector element that is used to match Pods
that are managed by the Deployment. The existing deployment manifest
was missing a selector (and labels), and deploying it would result in:

> * spec.selector: Required value
> * spec.template.metadata.labels: Invalid value:
>   map[string]string(nil): `selector` does not match template `labels`

This commit adds a kustomization.yaml file with the "commonLabels"
element. This will cause kustomize to automatically configure the
labels on the Deployment, the Pod template, and the selector.